dev cinia (#109)
* feat: add detail address lab

* fix: change view to delete in Lab Service

* feat: add failed to checkout dialog

* feat: add lab rating

* fix: add store for feedback

* fix: order history in dashboard direct to order detail

* fix: take out forgot password when sign in

* feat: add page request test

* fix: address detail in find-lab

* feat: add image validation in add and detail service pages

* fix: change lockfileVersion and remove comments
