+++ src/views/Dashboard/Lab/Services/index.vue
@@ -61,16 +61,16 @@
                            @click="gotoDetails(item)"
                         >
                            <v-icon dark>
-                              mdi-eye
+                              mdi-pencil
                            </v-icon>
                         </v-btn>
                         <v-btn
                            elevation="0"
                            color="transparent"
-                           @click="gotoDetails(item)"
+                           @click="deleteService(item)"
                         >
                            <v-icon dark>
-                              mdi-pencil
+                              mdi-delete
                            </v-icon>
                         </v-btn>
                      </v-container>
@@ -87,6 +87,7 @@
 <script>
 import DataTable from '@/components/DataTable'
 import SearchBar from '@/components/DataTable/SearchBar'
+import { deleteService } from '@/lib/polkadotProvider/command/services'
 import { mapGetters } from 'vuex'
 
 export default {
@@ -108,6 +109,8 @@ export default {
   }),
   computed:{
    ...mapGetters({
+      api: 'substrate/getAPI',
+      pair: 'substrate/wallet',
       labAccount: 'substrate/labAccount',
    }),
   },
@@ -116,9 +119,31 @@ export default {
       this.search = val
    },
    gotoDetails(item){
+      
       this.$router.push({ name: 'lab-dashboard-services-detail', params: { item: item }})
    },
-  },
+   getImageLink(val){
+      if(val && val != ""){
+         return val
+      }
+      return "https://ipfs.io/ipfs/QmaGr6N6vdcS13xBUT4hK8mr7uxCJc7k65Hp9tyTkvxfEr"
+   },
+   async deleteService(item) {
+      const isConfirmed = confirm("Are you sure you want to delete this service?")
+      if(isConfirmed)  {
+         this.isLoading = true
+         await deleteService(
+            this.api,
+            this.pair,
+            item.id,
+            () => {
+            this.$router.push('/lab/services')
+            this.isLoading = false
+            })
+         }
+         return
+      },
+  }
 }
